MainSceneAlex, new gameobject selection
So, basically now we can avoid using arraylists manually, and we can just add public gameobjects in the inspector.
The rule of thumb: number of this objects=number of second array of objects(they on the same script) and they all equal to the number of spawners  of this type.For example if i use 2 spawners for small objects, I need to use 2 equal gameobjects in the script , which you attach to these spawners. If you add new 3 spawners for medium objects, you attach this script to all of them , and put inside of each 3+3 equal game objects